Refactors evaluator and executor setup
Moves the evaluator logic from `radiate-engines` to `radiate-core` to promote reusability and decoupling.

Introduces a `FitnessEvaluator` struct that encapsulates the evaluation logic.
This allows for more flexible configuration of the engine,
including the ability to specify different executors for various steps.

Removes `SequentialEvaluator` and `WorkerPoolEvaluator` in favor of the single `FitnessEvaluator`.