Refactors input conversion to input transformation
Replaces the term "convert" with "transform" for input processing.
This change aims to provide a more accurate and descriptive term
for the operation performed on engine inputs, clarifying their role
in transforming data for use within the system.

The conversion logic is encapsulated within a new `InputTransform` trait
and used when selecting, altering and limiting the population.

This change improves code clarity and maintainability by renaming the
`InputConverter` trait to `InputTransform` and updating the
corresponding methods. Also adds py_alter and py_select functions to
allow for the selection and altering of populations from python code.