Improves graph codec and evaluator
Adds the ability to limit the maximum number of nodes in a graph.
This prevents the graph from growing indefinitely, which can improve
performance and stability.

Also adds the ability to reset the eval cache in the python binding.
This is required to reset the state of the graph between evaluations,
especially in recurrent graphs.

Refactors the graph evaluator to allow caching of the evaluation order,
input ranges, and outputs.
This dramatically increases performance on larger more complex graphs.