Refactors type handling and iteration for genomes
Updates type hints and iteration to be more flexible and consistent.
It allows chromosome and genotype constructors to accept any iterable of genes or chromosomes, respectively, instead of only lists.
Also updates the return type of gene_type methods to use the GeneType enum and adds a setter to Population to enable setting Phenotypes at a specific index.